feat: allow importing and publishing the library
This diff refactors the code such that it's possible to
import the IQB public library locally and also to publish
it using pypi idiomatically.

While there, add public API to the library `__init__.py`
so that we can change file names within the library source
code without breaking the library users' code.

While there, introduce a CI workflow so that we know
that new commits do not break the library.

We currently only have smoke tests but soon we're going to
introduce a `pytest` dependency and have proper tests.