fix: start addressing `ruff check --fix .` issues (#9)
This pull request starts addressing the issues surfacing when running:

```bash
uv run ruff check --fix .
```

I noticed that the current `ruff` configuration erred on the side of DRY
but I would rather avoid that for domain clarity, so I disabled it.

In a subsequent diff, I'll address more issues that I preferred not to
address now, lest we produce a large diff.