refactor!: use snake case for metric keys
The original format used strings with spaces. While this is ~okay
with Python, it creates lots of friction when using the same structures
in JavaScript.

In fact, when in JSON you have:

```JSON
{"download_speed_mbps": 11.4}
```

you can use it in JavaScript as:

```JavaScript
data.download_speed_mbps
```

whereas, when you have:

```JSON
{"download speed": 11.4}
```

you are forced to use:

```JavaScript
data["download speed"]
```

which is less convenient and less idiomatic.

Therefore, because IQB aims to be used from JavaScript, it seems
much better to err on the side of idiomaticity.

BREAKING CHANGE: use snake case for IQB metrics keys.