refactor: align percentile semantics across metrics
On page 19 of the [IQB report 2025](
https://www.measurementlab.net/publications/IQB_report_2025.pdf)
we read:

> IQB uses the 95th percentile of a given
> dataset to evaluate a given metric. In this
> context, the 95th percentile is the value
> below which 95% of the observed
> measurements fall, which effectively
> captures the upper bound of a typical user
> experience while excluding extreme
> outliers. For example, to assess whether a
> region meets the network tier’s packet loss
> criteria for high-quality gaming, IQB
> calculates the 95th percentile of packet
> loss measurements collected from users in
> that region. The value is then compared to
> the predefined threshold. 

The formula example (page 33) shows taking the 95p of
metrics without considerations regarding polarity.

However, to produce consistent results, we need to take
the polarity into account when taking the 95p.

Let us illustrate the polarity issue with an example. We
assume that the following holds in a specific ISP:

(i) the 95p of latency being 10 ms means that 95% of
samples have 10 ms or less

(ii) the 95p of download speed being 22 Mbit/s means that
95% of samples have 22 Mbit/s or less

Let us also assume that we're evaluating online gaming and
that online gaming typically needs:

(a) download speed >= 20 Mbit/s

(b) latency <= 15 ms

The latency percentile allows us to say that "most
samples" (95%) in the given ISP show a latency (10 ms
or better) lower than the required latency (15 ms).

The speed percentile allows us to say that "few samples"
(5%) in the given ISP show a download speed (22 Mbit/s
or better) greater than the required one (20 Mbit/s).

So, is online gaming possible in the ISP? The answer
seems nonconclusive because of the imbalance in the
samples we are considering and the different polarity
between latency (higher is worse) and speed (higher
is better).

On paper, a better solution is to say "okay, for the
speed, instead, we consider the 5p". Now, let us assume:

(iii) the 5p of download speed being 21 Mbit/s means that
5% of samples have 21 Mbit/s or less

Based on this we can say that 95% of users have 21 Mbit/s
or more. Now, it's possible to write a statement regarding
the download speed concerning "most samples".

This allows us to conclude "most samples indicate that
users can play online with this ISP".

How to translate this into the actual code?

**Approach I**: modify `cache.py` so that, when we request
`percentile=X` we actually take the complementary percentile
for latency and loss (or for download and upload speed).

This change aligns the polarity and allows us to answer
questions using uniform sample sizes. It is basically
equivalent to what we manually did above.

**Approach II**: swap the labels for latency and loss (or for
download speed and upload speed) when querying BigQuery.

This means `cache.py` uniformly accesses `percentile=X` with
the understanding that 2/4 of the labels are swapped.

The second approach is more robust because it guarantees that,
if `percentile=X` is there then also the complementary
percentile is there. In both cases, people reading the code
will need to be aware of the polarity anyway.

Based on a discussion with @sermpezis, I am going to swap the
labels for latency and packet loss. The actual swapping operation
is anyway irrelevant (it's mostly a matter of convention) and
what matters is that we're aligning the polarity.

The meaning of the approach we are choosing is that 95p is the
cutoff where 95% of users have worse performance for the definition
of worse that is implied by the metric (e.g., lower speed or
higher latency). Obviously, it also holds the opposite, that is,
that 5% of users have better performance.

In conclusion, the implemented change aligns the sample size so
that the same percentile label picked up by `cache.py` allows
to make comparable statements with respect to better/worse.