refactor(data): move JSON files to ./data/cache/v0
This refactor addresses friction I felt while I was working on
https://github.com/m-lab/iqb/pull/28.

Having the data at the same level of the scripts used to generate
them significantly makes it harder to inspect what changed.

Additionally, scoping to a specific data version allows us to
keep the current files around for as long as we need them, but
allows incrementally extending the data fetching pipeline.