fix(data): use `measurement-lab` project
The [documentation](https://www.measurementlab.net/data/docs/bq/quickstart/)
explicitly mentions `measurement-lab` as the project.

I was previously using the `mlab-sandbox` project instead. And there was
a TODO issue assigned to me to sort this out.

FWIW, re-running the pipeline *after* changing the project does not
yield different results in terms of number of tests per country. Therefore,
I suspect the impact of this change is epsilon. Yet, let us follow
the documented approach.

As noted previously, yet, BigQuery being distributed and quantiles being
approximate cause lots of churn in the quantile values when re-running
queries, so I am not committing the changes to avoid creating unnecessary
increases in the overall size of this repository.