refactor(data): use IQBPipeline to query
This diff changes `data/run_query.py` to use the `IQBPipeline` to query
for data and saves v1 cached data as a side effect.

Additionally, we continue to generate and update v0 data as part of
running the pipeline. In fact, the rest of the pipeline has not been
modified and only `data/run_query.py` internals have changed.

This diff opens up the way for stopping to use v0 data and just keep it
around as golden files for testing. On this note, add `.gitignore` rules
to `./data/cache` so that only the files currently part of the hierarchy
are kept around as reference and anything else is instead ignored.