feat(iqb/cache): implement reading parquet files
This diff adds support for reading parquet files efficiently
using filters and column projections.

The resulting API is a low-level API for accessing the data, which
seems handy because it allows for data exploration.

We will iterate over this API and build more user friendly APIs
for interacting with the available data.