refactor(pipeline)!: factor BigQuery-to-Parquet code
This commit separates the bigquery proper code and the IQB
specific logic for fetching data from bigquery.

The end result is this:

- `bqpq.py` (BigQuery to Parquet) contains generic code for
streaming BigQuery results to Parquet files

- `__init__.py` contains `IQBPipeline` that now uses the
code in `bqpq.py` to stream data.

As a result, files are smaller, tests are easier to read and
update, and more focused on specific units.

BREAKING CHANGE:

1. The `IQBPipeline.__init__` `project_id` argument has been
renamed to `project` for consistency with BigQuery

2. `IQBPipeline.get_cache_entry` now requires kwargs (for
safety, since they well all strings -- easy to make mistakes!)

3. `IQBPipeline.execute_query` also requires kwargs now

4. `QueryResult` has been renamed `PipelineBQPQQueryResult`
to make it crystal clear what it belongs to

5. `ParsedTemplateName` has been renamed `PipelineCacheTemplateName`
for the same reason as above (make it super clear)