feat(pipeline): introduce the dataset concept (#59)
We're going to use a flat `v1` cache where all the available datasets
(m-lab or not) are in the same directory.

To start preparing the ground for this, I need to conceptually replace
the `template` name in pipeline functions to a more generic concept of
`dataset_name`.

Additionally, since we don't know a priori whether each dataset will
consist of download and upload (they may or may not), we make it more
generic and define constructors that are specific to the dataset and
return the right string.

In turn, this allows to reduce the coupling between the pipeline, which
dictates the on-disk data format, and the cache, which uses
pipeline-provided abstractions for reading the cache.

Subsequent pull requests will make use of this change, in fact, to
decrease the coupling and separate responsibilities.