refactor(cache): separate mlab and generic code
This diff extracts mlab-specific code from the cache package
and creates a new module named `mlab.py` to contain it.

There are little to none functional changes here. The main
change is a slightly different exception wording when the
cache entry is missing.

While there, ensure we have better testing.