doc(analysis): document the pandas based API
This diff refactors the analysis notebook to illustrate how
a developer can obtain the data as DataFrames.

It also shows how to convert the obtained DataFrames to the data
format expected by the IQB calculator.

The availability of DataFrames allows for additional filtering
and exploration while working with the notebooks.