fix(queries): allow for geographical navigation
To navigate from country to subdivision1 to city we need to
have both country and subdivision1 inside city, otherwise it's
not possible to filter the more-fine-grained data frame.

For example, we may start with Italy. If we want to focus
on, say, Liguria, we need to have both Italy and Liguria in
the subdivision1 parquet table. Then, to go from that to
Italy/Liguria/Genoa we need both Italy and Liguria inside
of the city table, to be able to filter.

So, to address this issue, with this commit, I have added
the subdivision1 to the city-level queries.