feat: Update prototype UI with map functionality  (#83)
This PR updates the prototype with changes demoed at the Barcelona 2025
workshop:

- Adds map functionality 

- Modifies sunburst design with reusable component 

Note: This version contains hardcoded values for percentiles as well as
configurations for zoom in per country. These will be updated with the
next version of the prototype to be dynamically populated.

While there, move rather than copying `./cache/v0`, which was unused by
the library, so keep the diff ~small.

---------

Co-authored-by: Simone Basso <bassosimone@gmail.com>