feat(library): add ghremote package
The ghremote package implements `pipeline.RemoteCache` using
GitHub releases to publish files.

The `pipeline.RemoteCache` facility was introduced in
https://github.com/m-lab/iqb/pull/85.

The implementation is at `ghremote.IQBGitHubRemoteCache`.

We derived the implementation from already-existing code living
inside the `./data/ghcache.py` file.

Using GitHub releases to publish datasets is meant as an interim
solution. However, we are also going to implement a better approach
for datasets, possibly based on GCS buckets. To this end, we need
to refactor the existing approach to create the facilities for
making the GCS-based approach possible. Hence, this diff, so that
we can slowly move towards GCS-based publishing of datasets.

This diff adds the basic functionality and tests, along with minor
changes and tweaks in the rest of the codebase (boyscout rule).

We integrated the code added by this PR and verified it is working
as intended in https://github.com/m-lab/iqb/pull/87.