refactor(pipeline)!: pass remote_cache to ctors
This diff modifies the way in which we pass the remote_cache
kwarg to the pipeline. Rather than passing it to the method
that syncs the cache, let's use the constructor.

This feels more in line with object-oriented principles.

BREAKING CHANGE: the `remote_cache` key-value argument is now
passed to constructors instead of being passed to the method
getting specific cache entries.