feat(pipeline)!: introduce file-based locking
We need to improve our concurrency story, especially considering
that the code may run in a service that may fetch new data.

The first step to ensure safety (implemented by this diff) is
implementing mutual exclusion when writing.

To this end, we use portable, advisory file locking.

While working on this diff, I also realized that part of the
`pipeline` code (and namely `PipelineCacheManager`) is lazy and
part (`IQBPipeline`) is not lazy with `fetch_if_missing`.

This design makes it problematic to fit a single locking schema
into the model. So, we remove the non-lazy part and expose to
the user the ability to `lock()` and `sync()` a `PipelineCacheEntry`.

This means that the user can also decide to avoid locking, which
makes sense when one is using this library for research.

While there, also expose `PipelineCacheEntry.exists()` so we don't
need to roll out manually code that checks for the existence of both
files, and which is more robust in case we add files to an entry.

BREAKING CHANGE: the `fetch_if_missing` argument has been removed
and now you need to call `.sync()` explicitly to sync.