fix(pipeline): ensure eventual consistency on Unix
This diff modifies the code that writes files to provide eventual
consistency on Unix, by using a temporary directory and only
moving the file in place when done writing it.

Coupled with advisory file locking for mutually excluding writers,
this strategy ensures we can have server-side code fetching and
caching data on demand (up until some reasonable maximum storage
size, or considering a reasonable eviction policy).

With Windows, this is not possible, but we're not going to run
a server on Windows for now, so we're fine.

With this change, the user story of making it possible to have some
degree of concurrent requests for data is now complete.