fix(pqread.py): allow subdivision1 filtering
This diff adds support for filtering for subdivision1 and
extends the tests to ensure we're testing for it.

While there, notice that the test suite was not adequately
testing filtering by columns (i.e., ensuring that we can get
the columns we want), thus introduce a fixture file
containing expected columns, to improve the tests quality.