feat(pipeline): improve operator experience
Add a progress bar letting the operator know the progress
of the BigQuery job and results download.

While there, refactor logging to avoid using the `root` logger
and use lazy parameter binding such that we do not perform
string interpolation unless there is a real need.

We're using `tqdm` to print the progress bar.