feat(cache): pass cache_manager to IQBCache
This diff adds a new kwarg to the constructor of IQBCache so that
we can configure a remote cache manager.

In turn, this allows people using the IQBCache to take advantage of
remotely cached queries and fetch them on demand.

While there, notice that `MLabCacheReader` is a misnomer because
we will soon be able to write the cache as a side effect, so reduce
the amount of surprise by renaming it to `MLabCacheManager`.