feat(cache)!: make read_cache_entry lazy (#100)
This commit main job is to make `read_cache_entry` lazy until we need to
read the data. This change allows for better usage of the bandwidth and
disk resources because we don't fetch what we don't need to read.

Additional changes in this diff:

1. Sync on read with per-entry lock to avoid concurrent writers. This
change ensures that concurrent writes do not mess up with the cached
files, thus making the codebase more robust.

2. Relax granularity validation to allow broader queries. We were only
checking when reading the pair anyway and the checks were quite dumb.
So, let us relax the checks and collect a bit more data about usage
before introducing something better.

3. Adjust sync behavior when no syncers exist; raise on missing files.
With this change, we preserve the previous behavior where a cache with
no syncers doesn't trigger exceptions (desirable!).

4. Add docs for lazy reads and potential exceptions. This change
clarifies how the API works.

5. update tests for new lazy/sync semantics. Required to merge.

6. Document potential improvements. Add TODOs about potential codebase
improvements and small bugs.

BREAKING CHANGE: Reading a cache entry at the `iqb.cache` level is now
lazy until we need the data.