fix: prefer os.replace to shutil.move
The original motivation for this diff is to avoid deleting the
file in `ghremote` before downloading the new version.

We want to avoid this because it creates a short time window in
which readers will fail because the file is missing.

So, here's a better strategy, where we use a temporary directory
in the same filesystem and `os.replace`.

On Unix, this will work very well because our new temporary directory
policy guarantees we are in the same filesystem.

On Windows, I verified that Arrow does not use the
`FILE_SHARE_DELETE` attribute, thus, if we write while
someone is reading, we see a `PermissionError`. Yet,
for now, this is just a case of YAGNI.

Once the `ghremote` issue has been fixed, we applied the same
pattern to other code location using `shutil.move`.