fix: prefer os.replace to shutil.move (#102)
The original motivation for this diff is to avoid deleting the file in
`ghremote` before redownloading.

We want to avoid this because it creates a short time window in which
readers will fail because the file is missing.

So, here's a better strategy, where we use a temporary directory in the
same filesystem and `os.replace`.

On Unix, this will work very well because Unix is designed to do this
atomically within the same filesystem.

On Windows, I verified that Arrow does not use the `FILE_SHARE_DELETE`
attribute, thus, if we write while someone is reading, we see a
`PermissionError`. Yet, for now, this is just a case of YAGNI: we don't
use Windows.

Hence, the approach is to document the potential issue and move on.

Once the `ghremote` issue has been fixed, we applied the same pattern to
other code locations using `shutil.move`.