feat(scripting): add exception interceptor
This functionality allows to intercept and log exceptions
without too many try..catch and unwinding.

It also allow to get the proper sys.exit exitcode.

I tested this functionality briefly in an integration
branch, and it seems to help to have better UX while
at the same time having a reasonable DX.