cleanup(data): remove cached entries (#120)
The remote-cache code does not handle magic cached entries in the
repository. We've already published the related files such that it's
possible to retrieve them from GitHub.

What remains to be done (and we do in this commit) is to move the cached
files elsewhere. Specifically:

1. we use 2024-10 data as test fixtures for library

2. we use 2025-10 data as test fixtures for analysis (with the added
benefit of not requiring network I/O for these tests)

This removes testing of 2025-10 for the library but honestly it is not a
problem. Originally, we added those tests when migrating from JSON to
Parquet, however now we don't need to keep testing with both periods of
time.

To make this work, BTW, we need to manually add two more manifests for
the related directories.

While there, make sure we centralize test fixtures.