doc(data): update README.md (#121)
Explain that `./data` does not contain static data anymore and is now
basically a workspace with scripts that we've not properly integrated
into IQB yet. While there, also update the related READMEs to mention
recent changes that caused us to stop using static data in `./data` and
which ultimately changed the semantics of `./data`.